% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force search.}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A a data.frame x with two variables v and w.}

\item{W}{The knapsack size.}
}
\value{
A list with the maximum knapsack value and which elements.
}
\description{
Creating Brute force search algorithm to solve knapsack problem.
}
\details{
Input takes a data.frame x with two variables v and w and returns the maximum knapsack value and which elements (rows in the data.frame).
}
