% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_objects.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Knapsack objects with their weight(w) and value(v).}
\format{
A data frame with 2000 rows and 2 variables:
}
\description{
A data set containing the item weight and item value of almost 2,000
knapsack objects.
}
